Organize simulation into modules
